carnest-backend/requirements.txt
Django
django-rest-framework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
asgiref
django-cors-headers
django-dotenv
PyJWT
pytz
sqlparse
tzdata
python-dotenv
requests
Pillow
channels
channels_redis
daphne
twisted[http2,tls]
boto3
django-storages
